% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzePopulation.R
\name{analizePopulation}
\alias{analizePopulation}
\title{Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot}
\usage{
analizePopulation(
  envir,
  methylationData,
  slidingWindowSize,
  betaSuperiorThresholds,
  betaInferiorThresholds,
  sampleSheet,
  betaMedians,
  bonferroniThreshold = 0.05,
  probeFeatures
)
}
\arguments{
\item{envir}{semseekere working infos}

\item{methylationData}{whole matrix of data to analyze.}

\item{slidingWindowSize}{size of the sliding widows to compute epilesions
default 11 probes.}

\item{betaSuperiorThresholds}{data frame to select, from the sample sheet,
samples to use as control as study population and as refereces two vectors
within the first vector the names of the selection colum and tha second
vector the study population selector,}

\item{betaInferiorThresholds}{name of samplesheet's column to use as control
population selector followed by selection value,}

\item{sampleSheet}{name of samplesheet's column to use as control population
selector followed by selection value,}

\item{betaMedians}{name of samplesheet's column to use as control population
selector followed by selection value,}

\item{bonferroniThreshold}{threshold to define which pValue accept for}

\item{probeFeatures}{probes detail from 27 to EPIC illumina dataset
lesions definition}
}
\value{
files into the result folder with pivot table and bedgraph.
}
\description{
Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot
}
