% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_population.R
\name{analyze_population}
\alias{analyze_population}
\title{Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot}
\usage{
analyze_population(
  signal_data,
  sample_sheet,
  signal_thresholds,
  probe_features
)
}
\arguments{
\item{signal_data}{whole matrix of data to analyze.}

\item{sample_sheet}{name of samplesheet's column to use as control population
selector followed by selection value,}

\item{signal_thresholds}{thresholds defined to calculate epimutations}

\item{probe_features}{probe_features detail from 27 to EPIC illumina dataset}
}
\value{
files into the result folder with pivot table and bedgraph.
}
\description{
Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot
}
