% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSingleSample.R
\name{analyzeSingleSample}
\alias{analyzeSingleSample}
\title{analyzeSingleSample}
\usage{
analyzeSingleSample(
  values,
  slidingWindowSize,
  thresholds,
  figure,
  sampleDetail,
  bonferroniThreshold = 0.05,
  probeFeatures
)
}
\arguments{
\item{values}{values of methylation}

\item{slidingWindowSize}{size of window sliding to calculate hypergeometric}

\item{thresholds}{threshold to use for comparison}

\item{bonferroniThreshold}{bonferroni threshold to validate pValue}

\item{probeFeatures}{probes features probe, chr, start,end}

\item{sampleDetail$Sample_ID}{name of sample to analyze}

\item{subFileExtension}{extension to pre pend to file name}
}
\value{
list of lesion count  and probes count
}
\description{
analyzeSingleSample
}
