% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSingleSampleBothThresholds.R
\name{analyzeSingleSampleBothThresholds}
\alias{analyzeSingleSampleBothThresholds}
\title{analyzeSingleSampleBothThresholds}
\usage{
analyzeSingleSampleBothThresholds(
  values,
  resultFolder,
  betaSuperiorThresholds,
  betaInferiorThresholds,
  sampleName,
  probeFeatures
)
}
\arguments{
\item{values}{values of methylation}

\item{resultFolder}{folder to store result files}

\item{betaSuperiorThresholds}{threshold for hyper methylation}

\item{betaInferiorThresholds}{threshold for hypo methylation}

\item{sampleName}{name of sample to analyze}

\item{probeFeatures}{probes features probe, chr, start,end}

\item{slidingWindowSize}{size of window sliding to calculate hypergeometric}

\item{subFileExtension}{extension to pre pend to file name}
}
\value{
list of lesion count  and probes count
}
\description{
analyzeSingleSampleBothThresholds
}
