% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_bed.R
\name{annotate_bed}
\alias{annotate_bed}
\title{takes a bed and its location (build with the details of population and genomic area)
and annotate with detail about genomic area, the output file is only for the arean and specified subarea}
\usage{
annotate_bed(sample_groups, figures, markers, area, subarea)
}
\arguments{
\item{sample_groups}{vector of population to cycle with to build the folder path}

\item{figures}{vector of hyper /hypo to use to build the folder path}

\item{markers}{vector of lesions/mutations to use to build the folder path}

\item{area}{vector of genomic area to cycle and group the annotated data}

\item{subarea}{subarea of the area to annotate..}
}
\value{
original bed with genomic area infos
}
\description{
takes a bed and its location (build with the details of population and genomic area)
and annotate with detail about genomic area, the output file is only for the arean and specified subarea
}
