% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateBed.R
\name{annotateBed}
\alias{annotateBed}
\title{Title}
\usage{
annotateBed(
  populations,
  figures,
  anomalies,
  groups,
  probesPrefix,
  columnLabel,
  groupingColumnLabel,
  resultFolder
)
}
\arguments{
\item{resultFolder}{}
}
\value{

}
\description{
Title
}
