% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_quantreg_summary.R
\name{assoc_exact_pvalue}
\alias{assoc_exact_pvalue}
\title{Quantile regression result value, confidence interval and p.value}
\usage{
assoc_exact_pvalue(permutation_vector, statistic_parameter, conf.level)
}
\arguments{
\item{permutation_vector}{vector of permutation statistc signal regression}

\item{statistic_parameter}{signal regression}

\item{conf.level}{confidence intervals alpha level}
}
\value{
ci and p.value and pvalue_limit
}
\description{
Quantile regression result value, confidence interval and p.value
}
