% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_single_sample.R
\name{beta_single_sample}
\alias{beta_single_sample}
\title{beta_single_sample}
\usage{
beta_single_sample(values, sample_detail, probe_features)
}
\arguments{
\item{values}{beta vaòues}

\item{sample_detail}{detais of sample}

\item{probe_features}{annotation probe}
}
\value{
beta mean
}
\description{
beta_single_sample
}
