% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data_set_from_geo.R
\name{build_data_set_from_geo}
\alias{build_data_set_from_geo}
\title{build_data_set_from_geo}
\usage{
build_data_set_from_geo(GEOgse, downloadFiles = 0, result_folder, ...)
}
\arguments{
\item{GEOgse}{geo accession dataset identification}

\item{downloadFiles}{0 means download all files from Gene Expression Ombibus (GEO),
different than zero means how many download}

\item{result_folder}{where sample sheet and files will be saved}
}
\value{
samplesheet, and sample's file saved and samplesheet csv
}
\description{
build_data_set_from_geo
}
\examples{
result_dir <- tempdir()
\donttest{
build_data_set_from_geo(
  GEOgse        = "GSE55763",
  downloadFiles = 1,
  result_folder = tempdir()
)
}
}
