% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot}
\usage{
calculate(
  sampleSheet,
  methylationData,
  probeFeatures,
  resultFolder,
  logFolder,
  slidingWindowSize = 11,
  popControlSelector,
  popStudyDefinition,
  bonferroniThreshold = 0.05
)
}
\arguments{
\item{resultFolder}{folder to save computed epimutations and bedgraphs files.}

\item{logFolder}{folder to output log}

\item{slidingWindowSize}{size of the sliding widows to compute epilesions
default 11 probes.}

\item{popControlSelector}{name of samplesheet's column to use as control
population selector followed by selection value,}

\item{popStudyDefinition}{data frame to select, from the sample sheet,
samples to use as control as study population and as refereces two vectors
within the first vector the names of the selection colum and tha second
vector the study population selector,}

\item{bonferroniThreshold}{= 0.05 #threshold to define which pValue
accept for lesions definition}
}
\value{
files into the result folder with pivot table and bedgraph.
}
\description{
Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot
}
\examples{
calculate(
slidingWindowSize = 11,
resultFolder = resultFolderValue,
logFolder = logFolderValue,
popStudyDefinition  = ,
popControlSelector = "Case_Ctrl == 0",
bonferroniThreshold = 0.05
)
}
