% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_bootstrap_model.R
\name{compute_mean_delta_boot}
\alias{compute_mean_delta_boot}
\title{Title}
\usage{
compute_mean_delta_boot(sig.formula, df, shuffle = FALSE)
}
\arguments{
\item{sig.formula}{formula to apply}

\item{df}{dataframe to use}

\item{tau}{tau at which apply the quantile regression}

\item{lqm_control}{specification of the lqmm package}
}
\description{
Title
}
