% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_permutation_model.R
\name{compute_quantile_delta_permutation}
\alias{compute_quantile_delta_permutation}
\title{Title}
\usage{
compute_quantile_delta_permutation(
  sig.formula,
  df,
  shuffle = FALSE,
  quantile = 0.5
)
}
\arguments{
\item{sig.formula}{formula to apply}

\item{df}{dataframe to use}

\item{tau}{tau at which apply the quantile regression}

\item{lqm_control}{specification of the lqmm package}
}
\description{
Title
}
