% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantreg_model.R
\name{compute_quantreg_beta_boot_np}
\alias{compute_quantreg_beta_boot_np}
\title{Title}
\usage{
compute_quantreg_beta_boot_np(sig.formula, df, tau, lqm_control)
}
\arguments{
\item{lqm_control}{}
}
\description{
Title
}
