% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heatmap.R
\name{create_heatmap}
\alias{create_heatmap}
\title{create_heatmap load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot}
\usage{
create_heatmap(inputBedDataFrame, anomalies, file_prefix, groupColumnLabels)
}
\arguments{
\item{inputBedDataFrame}{data frame to chart}

\item{anomalies}{vector of anomalies to manage}

\item{file_prefix}{main genomic area to char eg: gene}

\item{groupColumnLabels}{positions of the group coplumn id}
}
\value{
list of pivot by column identified with file_prefix and by Sample
}
\description{
create_heatmap load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot
}
