% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_position_pivots.R
\name{create_position_pivots}
\alias{create_position_pivots}
\title{create pivot from samples' single bed for the PROBE (WHOLE)}
\usage{
create_position_pivots(population, keys)
}
\arguments{
\item{population}{population data frame to use for creating pivot}

\item{keys}{markers and figure to create pivot of}
}
\description{
create pivot from samples' single bed for the PROBE (WHOLE)
}
