% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createChartFromMultipleBedGenericPerRegion.R
\name{createChartFromMultipleBedGenericPerRegion}
\alias{createChartFromMultipleBedGenericPerRegion}
\title{createChartFromMultipleBedGenericPerRegion load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot}
\usage{
createChartFromMultipleBedGenericPerRegion(
  resultFolder,
  populations,
  figures,
  anomalies,
  subGroups,
  probesPrefix,
  mainGroupLabel,
  subGroupLabel
)
}
\arguments{
\item{resultFolder}{folder as root for bedfiles organized per anomaly}

\item{populations}{vectors of populations name used to group populations
for differential analysis}

\item{figures}{vectors of figure to identify lesions as HYPER or HYPO}

\item{anomalies}{vector of anomalies MUTATIONS or LESIONS}

\item{subGroups}{other grouping column over main group}

\item{probesPrefix}{definition of prefix to use probes to load}

\item{mainGroupLabel}{name of column to group the data set}

\item{subGroupLabel}{name of column to sub group the data set}
}
\value{
list of pivot by column identified with mainGroupLabel and by Sample
}
\description{
createChartFromMultipleBedGenericPerRegion load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot
}
