% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHeatmap.R
\name{createHeatmap}
\alias{createHeatmap}
\title{createHeatmap load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot}
\usage{
createHeatmap(inputBedDataFrame, anomalies, groupLabel, groupColumnIDs)
}
\arguments{
\item{inputBedDataFrame}{data frame to chart}

\item{groupColumnIDs}{}
}
\value{
list of pivot by column identified with groupLabel and by Sample
}
\description{
createHeatmap load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot
}
