% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPivotResultFromMultipleBed.R
\name{createPivotResultFromMultipleBed}
\alias{createPivotResultFromMultipleBed}
\title{createPivotResultFromMultipleBed load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot}
\usage{
createPivotResultFromMultipleBed(
  resultFolder,
  anomalyLabel,
  figureLable,
  probeFeatures
)
}
\arguments{
\item{resultFolder}{folder as root for bedfiles organized per anomaly}

\item{anomalyLabel}{anomaly definition used to lable folder and files eg
MUTATIONS, LESIONS}

\item{figureLable}{figure used to create the file HYPO HYPER}

\item{probeFeatures}{features of probe CHR and START and NAME}
}
\value{
list of pivot by CHR and by Gene
}
\description{
createPivotResultFromMultipleBed load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot
}
