% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPivotResultFromMultipleBedGeneric.R
\name{createPivotResultFromMultipleBedGeneric}
\alias{createPivotResultFromMultipleBedGeneric}
\title{createPivotResultFromMultipleBedGeneric load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot}
\usage{
createPivotResultFromMultipleBedGeneric(
  genomicAreaMain,
  genomicAreaSub,
  sampleSheet
)
}
\arguments{
\item{genomicAreaMain}{is the genomic area to investigate for example the gene}

\item{genomicAreaSub}{is the part of the genomic area main for example the Body of the gene}

\item{sampleSheet}{sample sheet to use}
}
\value{
list of pivot by column identified with columnLabel and by Sample
}
\description{
createPivotResultFromMultipleBedGeneric load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot
}
