% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummaryExcelFromCumulativeBedFile.R
\name{createSummaryExcelFromCumulativeBedFile}
\alias{createSummaryExcelFromCumulativeBedFile}
\title{createSummaryExcelFromCumulativeBedFile}
\usage{
createSummaryExcelFromCumulativeBedFile(
  resultFolder,
  probeFeaturesOriginal,
  sampleSheet
)
}
\arguments{
\item{sampleSheet}{}
}
\value{

}
\description{
createSummaryExcelFromCumulativeBedFile
}
