% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpSampleAsBedFile.R
\name{dump_sample_as_bed_file}
\alias{dump_sample_as_bed_file}
\title{given data and colnames dump as bed file}
\usage{
dump_sample_as_bed_file(dataToDump, fileName)
}
\arguments{
\item{dataToDump}{data frame to dump into bed file with CHR, START, END}

\item{fileName}{name of the file to save data in}
}
\value{
nothing
}
\description{
given data and colnames dump as bed file
}
