% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpSampleAsBedFile.R
\name{dumpSampleAsBedFile}
\alias{dumpSampleAsBedFile}
\title{given data and colnames dump as bed file}
\usage{
dumpSampleAsBedFile(
  dataToDump,
  fileExtension,
  resultFolder,
  resultSubFolder,
  sampleName,
  multipleFileColNames
)
}
\arguments{
\item{dataToDump}{data frame to dump into bed file with CHR, START, END}

\item{fileExtension}{extension to give to written file}

\item{resultFolder}{root folder to save dumped files}

\item{resultSubFolder}{folder under the root result folder to group files}

\item{sampleName}{name of the sample to create the file for}

\item{multipleFileColNames}{names to use as header for the files dumped}
}
\value{
nothing
}
\description{
given data and colnames dump as bed file
}
