% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorative_analysis.R
\name{explorative_analysis}
\alias{explorative_analysis}
\title{Preliminary explorative analysys of the data}
\usage{
explorative_analysis(
  categorical_variables,
  numerical_variables,
  sample_sheet,
  signal_data,
  max_missed_sample_sheet = 0.3,
  max_missed_signal_data = 0.1,
  sample_id_column = "Sample_ID",
  exploration_phase = "0",
  sample_sheet_mapping = NULL,
  values_mapping = c(),
  result_folder,
  ...
)
}
\arguments{
\item{sample_sheet}{path to the sample sheet}

\item{signal_data}{path to the signal data}

\item{max_missed_sample_sheet}{max number of missing values in each sample sheet column}

\item{max_missed_signal_data}{max number of missing values in each signal data column and row}

\item{sample_id_column}{name of the columnwith sample ids}

\item{exploration_phase}{numerioc value to preserve histor of explorative analysis}

\item{sample_sheet_mapping}{rules to rename and remove columns}

\item{values_mapping}{rules to recode values and remove samples (source missed leave blank in the mapping file)}
}
\value{
reports of the explorative analysis and sample sheet, signal data cleaned
}
\description{
Preliminary explorative analysys of the data
}
