% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Calculate stochastic epi mutations from a methylation dataset as outcome report of pivot}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{a value.}

\item{b}{b value}
}
\value{
files into the result folder with pivot table and bedgraph.
}
\description{
Calculate stochastic epi mutations from a methylation dataset as outcome report of pivot
}
\examples{
fbind(
a = 11,
b = resultFolderValue
)
}
