% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ontology_analysis_rrvgo.R
\name{geneontology_analysis_rrvgo}
\alias{geneontology_analysis_rrvgo}
\title{geneontology_analysis_rrvgo}
\usage{
geneontology_analysis_rrvgo(resultFolder, fileName)
}
\arguments{
\item{resultFolder}{location of folder of semseeker result}

\item{fileName}{name of inference file to use with rrvgo}
}
\value{
null
}
\description{
geneontology_analysis_rrvgo
}
