% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLesions.R
\name{getLesions}
\alias{getLesions}
\title{for each sample given the mutations return the lesions}
\usage{
getLesions(
  mutationAnnotatedSorted,
  slidingWindowSize,
  sampleName,
  bonferroniThreshold = 0.05,
  probeFeatures
)
}
\arguments{
\item{mutationAnnotatedSorted}{dataframe of mutations annotated and sorted
per the input sample}

\item{slidingWindowSize}{window size to calculate the hypergeometric
distribution}

\item{sampleName}{name of the sample which working with}

\item{bonferroniThreshold}{accepted value to consider pValue valid}
}
\value{
data frame with lesioned/not lesioned probes identified by 1/0
}
\description{
for each sample given the mutations return the lesions
}
