% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutations.R
\name{getMutations}
\alias{getMutations}
\title{getMutations}
\usage{
getMutations(values, figure, thresholds, probeFeatures, sampleName)
}
\arguments{
\item{values}{values of methylation}

\item{thresholds}{threshold to use for comparison}

\item{probeFeatures}{probes features probe, chr, start,end}

\item{comparison}{function to use for compare}
}
\value{
mutations
}
\description{
getMutations
}
