% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis_phenolizer.R
\name{inference_pathway_analysis_phenolizer}
\alias{inference_pathway_analysis_phenolizer}
\title{inference_pathway_analysis_phenolizer calculate and att to sample sheet result
the number of mutations or the lesioned gene per each sample}
\usage{
inference_pathway_analysis_phenolizer(resultFolder, terms)
}
\arguments{
\item{resultFolder}{where semseeker result are saved}

\item{terms}{gene HP term to be searched}
}
\value{

}
\description{
inference_pathway_analysis_phenolizer calculate and att to sample sheet result
the number of mutations or the lesioned gene per each sample
}
