% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_env.R
\name{init_env}
\alias{init_env}
\title{init ssEnvonment}
\usage{
init_env(result_folder, maxResources = 90, ...)
}
\arguments{
\item{result_folder}{where result of semseeker will bestored}

\item{maxResources}{percentage of how many available cores will be used default 90 percent, rounded to the lowest integer}

\item{...}{other options to filter elaborations, including \code{parallel_strategy}
(parallelisation strategy for \pkg{future}: "sequential", "multisession",
"multicore", "cluster"; default "sequential")}
}
\value{
the working ssEnvonment
}
\description{
init ssEnvonment
}
