% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_build_data_set_from_geo.R
\name{io_build_data_set_from_geo}
\alias{io_build_data_set_from_geo}
\title{io_build_data_set_from_geo}
\usage{
io_build_data_set_from_geo(GEOgse, downloadFiles = 0, result_folder, ...)
}
\arguments{
\item{GEOgse}{geo accession dataset identification}

\item{downloadFiles}{0 means download all files from Gene Expression Omnibus (GEO),
different than zero means how many to download}

\item{result_folder}{where sample sheet and files will be saved}

\item{...}{additional arguments passed to internal helpers}
}
\value{
samplesheet, and sample's file saved and samplesheet csv
}
\description{
io_build_data_set_from_geo
}
\examples{
result_dir <- tempdir()
\dontrun{
io_build_data_set_from_geo(
  GEOgse        = "GSE55763",
  downloadFiles = 1,
  result_folder = tempdir()
)
}
}
