% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_dump_sample_as_bed_file.R
\name{io_dump_sample_as_bed_file}
\alias{io_dump_sample_as_bed_file}
\title{given data and colnames dump as bed file}
\usage{
io_dump_sample_as_bed_file(data_to_dump, fileName)
}
\arguments{
\item{data_to_dump}{data frame to dump into bed file with CHR, START, END}

\item{fileName}{name of the file to save data in}
}
\value{
nothing
}
\description{
given data and colnames dump as bed file
}
