% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_coord_input.R
\name{io_is_coord_format}
\alias{io_is_coord_format}
\title{Detect whether a data frame is in coordinate format (CHR + START columns).}
\usage{
io_is_coord_format(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
Logical scalar.
}
\description{
Detect whether a data frame is in coordinate format (CHR + START columns).
}
