% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot_marker_per_probe.R
\name{manhattan_plot_marker_per_probe}
\alias{manhattan_plot_marker_per_probe}
\title{Title manhattan_plot_marker_per_probe}
\usage{
manhattan_plot_marker_per_probe(
  probe_name = "cg11680158",
  max_sample = 0,
  min_sample = 0,
  min_signal_probe = 0,
  label_font_size = 3,
  hyper_color = "blue",
  hypo_color = "orange",
  non_outlier_color = "grey",
  limit_label_color = "blue",
  limit_line_color = "red",
  limit_line_color_median = "black",
  reference_samples_color = "cyan",
  show_labels = TRUE,
  result_folder = result_folder,
  maxResources = maxResources,
  parallel_strategy = parallel_strategy,
  ...
)
}
\description{
Title manhattan_plot_marker_per_probe
}
