% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot_marker_per_sample.R
\name{manhattan_plot_marker_per_sample}
\alias{manhattan_plot_marker_per_sample}
\title{Title manhattan_plot_marker_per_sample}
\usage{
manhattan_plot_marker_per_sample(
  sample_name = "NAME",
  probes_range = 1000:2000,
  hyper_color = "blue",
  hypo_color = "orange",
  non_outlier_color = "grey",
  limit_label_color = ssEnv$color_palette[1],
  result_folder = result_folder,
  maxResources = maxResources,
  parallel_strategy = parallel_strategy,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title manhattan_plot_marker_per_sample
}
