% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_bootstrap_model.R
\name{mean_bootstrap}
\alias{mean_bootstrap}
\title{Title}
\usage{
mean_bootstrap(
  family_test,
  sig.formula,
  tempDataFrame,
  independent_variable,
  boot_success,
  tests_count
)
}
\arguments{
\item{family_test}{family lqmm, mean}

\item{sig.formula}{formula of the model}

\item{tempDataFrame}{data}

\item{independent_variable}{name of regressor}

\item{boot_success}{number of success tests to calculate corrected confidence interval}

\item{tests_count}{count of total executed tests}
}
\description{
Title
}
