% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMultipleBed.R
\name{mergeMultipleBed}
\alias{mergeMultipleBed}
\title{Title}
\usage{
mergeMultipleBed(
  populations,
  figures,
  anomalies,
  fileExtension,
  resultFolder,
  multipleFileColNames
)
}
\arguments{
\item{multipleFileColNames}{}
}
\value{

}
\description{
Title
}
