% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_to_long_format.R
\name{pivot_to_long_format}
\alias{pivot_to_long_format}
\title{Get the pivot in long format instead of wide format}
\usage{
pivot_to_long_format(
  anomaly,
  figure,
  group,
  subgroup,
  phenotype_column,
  areas_selection = NULL
)
}
\arguments{
\item{phenotype_column}{column from the sample sheet to pair to each sample}

\item{areas_selection}{genomic area to select, if NULL all areas will be selected}

\item{area_pivot}{pivot dataframe to transform in long format}
}
\value{
the pivot in a long format of 3 columnns, the phontype column with name phenotype, the value of the anomaly and the area investigated
}
\description{
Get the pivot in long format instead of wide format
}
