% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_create_heatmap.R
\name{plot_create_heatmap}
\alias{plot_create_heatmap}
\title{plot_create_heatmap load the multiple bed resulting from
analysis organized into files and folders per marker and produce a pivot}
\usage{
plot_create_heatmap()
}
\value{
nothing
}
\description{
plot_create_heatmap load the multiple bed resulting from
analysis organized into files and folders per marker and produce a pivot
}
