% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROBES_Gene_TSS200}
\alias{PROBES_Gene_TSS200}
\title{PROBES_Gene_TSS200}
\format{
A data frame with five variables: \code{year}, \code{sex},
  \code{name}, \code{n} and \code{prop} (\code{n} divided by total number
  of applicants in that year, which means proportions are of people of
  that sex with that name born in that year).
}
\usage{
PROBES_Gene_TSS200
}
\description{
Full data set probes as defined by Illumina
}
\keyword{datasets}
