% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROBES_Island_S_Shore}
\alias{PROBES_Island_S_Shore}
\title{PROBES_Island_S_Shore}
\format{
A data frame with five variables: \code{year}, \code{sex},
  \code{name}, \code{n} and \code{prop} (\code{n} divided by total number
  of applicants in that year, which stats::means proportions are of people of
  that sex with that name born in that year).
}
\usage{
PROBES_Island_S_Shore
}
\description{
Full data set probes as defined by Illumina
}
\keyword{datasets}
