% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantreg_summary.R
\name{quantreg_summary}
\alias{quantreg_summary}
\title{Quantile regression result value, confidence interval and pvalue}
\usage{
quantreg_summary(
  boot_vector,
  estimate,
  working_data,
  sig.formula,
  tau,
  independent_variable,
  lqm_control
)
}
\arguments{
\item{boot_vector}{vector of boot statistc beta regression}

\item{estimate}{beta regression}

\item{working_data}{data to regress}

\item{sig.formula}{formula for model}

\item{tau}{quantile to regress at}

\item{independent_variable}{name of indenpendent variable}

\item{lqm_control}{controls of lqmm packages}
}
\description{
Quantile regression result value, confidence interval and pvalue
}
