% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/rangeBetaValuePerCHROverCHRAsNormalizedDistribution.R
\name{rangeBetaValuePerCHROverCHRAsNormalizedDistribution}
\alias{rangeBetaValuePerCHROverCHRAsNormalizedDistribution}
\title{Title}
\usage{
rangeBetaValuePerCHROverCHRAsNormalizedDistribution(
  populationMatrix,
  iqrTimes = 1,
  probeFeatures
)
}
\arguments{
\item{probeFeatures}{}
}
\value{

}
\description{
Title
}
