% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/rangebetaValuePerGeneOverGeneAsNormalizedDistribution.R
\name{rangeBetaValuePerGeneOverGeneAsNormalizedDistribution}
\alias{rangeBetaValuePerGeneOverGeneAsNormalizedDistribution}
\title{Title}
\usage{
rangeBetaValuePerGeneOverGeneAsNormalizedDistribution(
  populationMatrix,
  iqrTimes = 1,
  probeFeatures
)
}
\arguments{
\item{probeFeatures}{}
}
\value{

}
\description{
Title
}
