% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/rangeBetaValuePerProbeAsNormalizedDistribution.R
\name{rangeBetaValuePerProbeAsNormalizedDistribution}
\alias{rangeBetaValuePerProbeAsNormalizedDistribution}
\title{Nrmalize the methyation level distributing as a normalized distribution}
\usage{
rangeBetaValuePerProbeAsNormalizedDistribution(populationMatrix, iqrTimes = 3)
}
\arguments{
\item{populationMatrix}{matrix of methylation for the population under calculation}

\item{iqrTimes}{inter quartile ratio used to normalize}
}
\value{
methylation matrix as normalized distribution
}
\description{
Nrmalize the methyation level distributing as a normalized distribution
}
\examples{
rangeBetaValuePerProbeAsNormalizedDistribution(
populationMatrix = controlPopulationMatrix,
iqrTimes = 3
)
}
