% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMultipleBed.R
\name{readMultipleBed}
\alias{readMultipleBed}
\title{readMultipleBed}
\usage{
readMultipleBed(
  resultFolder,
  anomalyLabel,
  figureLable,
  probeFeatures,
  columnLabel,
  populationName,
  groupingColumnLabel
)
}
\arguments{
\item{resultFolder}{folder as root for bedfiles organized per anomaly}

\item{anomalyLabel}{anomaly definition used to lable folder and files eg
MUTATIONS, LESIONS}

\item{probeFeatures}{features of probe CHR and START and NAME}

\item{groupingColumnLabel}{}
}
\value{
list of pivot by column identified with columnLabel and by Sample
}
\description{
readMultipleBed
}
