% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionDifferentialAnalysis.R
\name{regionDifferentialAnalysis}
\alias{regionDifferentialAnalysis}
\title{Title}
\usage{
regionDifferentialAnalysis(
  resultFolder,
  populations,
  figures,
  anomalies,
  subGroups,
  probesPrefix,
  mainGroupLabel,
  subGroupLabel
)
}
\arguments{
\item{subGroupLabel}{}
}
\value{

}
\description{
Title
}
