% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionDifferentialAnalysisPerGenomicArea.R
\name{regionDifferentialAnalysisPerGenomicArea}
\alias{regionDifferentialAnalysisPerGenomicArea}
\title{Title}
\usage{
regionDifferentialAnalysisPerGenomicArea(
  resultFolder,
  populations,
  figures,
  anomalies,
  subGroups,
  probesPrefix,
  mainGroupLabel,
  subGroupLabel
)
}
\arguments{
\item{subGroupLabel}{}
}
\value{

}
\description{
Title
}
