% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_analyze_single_sample.R
\name{sem_analyze_single_sample}
\alias{sem_analyze_single_sample}
\title{sem_analyze_single_sample}
\usage{
sem_analyze_single_sample(values, thresholds, figure, sample_detail)
}
\arguments{
\item{values}{values of methylation}

\item{thresholds}{threshold to use for comparison}

\item{figure}{which figure of sample will be analysed: HYPO or HYPER}

\item{sample_detail}{details of the sample to analyze}
}
\value{
list of lesion count and probe_features count
}
\description{
sem_analyze_single_sample
}
