% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_single_sample.R
\name{signal_single_sample}
\alias{signal_single_sample}
\title{signal_single_sample}
\usage{
signal_single_sample(values, sample_detail, probe_features)
}
\arguments{
\item{values}{signal vaòues}

\item{sample_detail}{detais of sample}

\item{probe_features}{annotation probe}
}
\value{
signal mean
}
\description{
signal_single_sample
}
