% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_by_chr_and_start.R
\name{sort_by_chr_and_start}
\alias{sort_by_chr_and_start}
\title{sort the dataframe using CHR and START sorting column first for CHR and after
for START}
\usage{
sort_by_chr_and_start(dataframe)
}
\arguments{
\item{dataframe}{dataframe to be sorted}
}
\value{
sorted dataframe
}
\description{
sort the dataframe using CHR and START sorting column first for CHR and after
for START
}
