% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortByCHRandStart.R
\name{sortByCHRandSTART}
\alias{sortByCHRandSTART}
\title{sort the dataframe using CHR and START sorting column first for CHR and after
for START}
\usage{
sortByCHRandSTART(dataframe)
}
\arguments{
\item{dataframe}{dataframe to be sorted}
}
\value{
sorted dataframe
}
\description{
sort the dataframe using CHR and START sorting column first for CHR and after
for START
}
