% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortPivot.R
\name{sortPivot}
\alias{sortPivot}
\title{sort the pivot table per SAMPLENAME}
\usage{
sortPivot(pivotTable)
}
\arguments{
\item{pivotTable}{data to be sorted}
}
\value{
pivt table sorted
}
\description{
sort the pivot table per SAMPLENAME
}
