% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_test_match_order.R
\name{util_test_match_order}
\alias{util_test_match_order}
\title{Title}
\usage{
util_test_match_order(x, y)
}
\arguments{
\item{x}{vector to compare}

\item{y}{vector to compare}
}
\value{
true if the order matches otherwise is false
}
\description{
Title
}
