% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzePopulation.R
\name{analizePopulation}
\alias{analizePopulation}
\title{Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot}
\usage{
analizePopulation(
  populationMatrix,
  slidingWindowSize,
  resultFolder,
  logFolder,
  betaSuperiorThresholds,
  betaInferiorThresholds,
  sampleSheet,
  betaMedians,
  populationName,
  bonferroniThreshold = 0.05,
  probeFeatures
)
}
\arguments{
\item{populationMatrix}{size of the sliding widows to compute epilesions
default 11 probes.}

\item{slidingWindowSize}{size of the sliding widows to compute epilesions
default 11 probes.}

\item{resultFolder}{folder to save computed epimutations and bedgraphs files.}

\item{logFolder}{folder to output log}

\item{betaSuperiorThresholds}{data frame to select, from the sample sheet,
samples to use as control as study population and as refereces two vectors
within the first vector the names of the selection colum and tha second
vector the study population selector,}

\item{betaInferiorThresholds}{name of samplesheet's column to use as control
population selector followed by selection value,}

\item{sampleSheet}{name of samplesheet's column to use as control population
selector followed by selection value,}

\item{betaMedians}{name of samplesheet's column to use as control population
selector followed by selection value,}

\item{populationName}{name of samplesheet's column to use as control
population selector followed by selection value,}

\item{bonferroniThreshold}{threshold to define which pValue accept for}

\item{probeFeatures}{probes detail from 27 to EPIC illumina dataset
lesions definition}
}
\value{
files into the result folder with pivot table and bedgraph.
}
\description{
Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot
}
