% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_population.R
\name{analyze_population}
\alias{analyze_population}
\title{Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot}
\usage{
analyze_population(
  methylation_data,
  sliding_window_size,
  sample_sheet,
  beta_thresholds,
  bonferroni_threshold = 0.05,
  probe_features
)
}
\arguments{
\item{methylation_data}{whole matrix of data to analyze.}

\item{sliding_window_size}{size of the sliding widows to compute epilesions
default 11 probe_features.}

\item{sample_sheet}{name of samplesheet's column to use as control population
selector followed by selection value,}

\item{beta_thresholds}{thresholds defined to calculate epimutations
lesions definition}

\item{bonferroni_threshold}{threshold to define which pValue accept for}

\item{probe_features}{probe_features detail from 27 to EPIC illumina dataset}
}
\value{
files into the result folder with pivot table and bedgraph.
}
\description{
Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot
}
