% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_single_sample.R
\name{analyze_single_sample}
\alias{analyze_single_sample}
\title{analyze_single_sample}
\usage{
analyze_single_sample(
  values,
  sliding_window_size,
  thresholds,
  figure,
  sample_detail,
  bonferroni_threshold = 0.05,
  probe_features
)
}
\arguments{
\item{values}{values of methylation}

\item{sliding_window_size}{size of window sliding to calculate hypergeometric}

\item{thresholds}{threshold to use for comparison}

\item{figure}{which figure's of sasmple will be analized HYPO or HYPER}

\item{sample_detail}{details of the sample to analyze}

\item{bonferroni_threshold}{bonferroni threshold to validate pValue}

\item{probe_features}{probe_features details to be used}
}
\value{
list of lesion count  and probe_features count
}
\description{
analyze_single_sample
}
