% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeSingleSample.R
\name{analyzeSingleSample}
\alias{analyzeSingleSample}
\title{Analyze a single sample}
\usage{
analyzeSingleSample(
  values,
  slidingWindowSize,
  resultFolder,
  thresholds,
  comparison,
  sampleName,
  subFileExtension,
  bonferroniThreshold = 0.05,
  probeFeatures
)
}
\arguments{
\item{values}{values of methylation}

\item{slidingWindowSize}{size of window sliding to calculate hypergeometric}

\item{resultFolder}{folder to store result files}

\item{thresholds}{threshold to use for comparison}

\item{comparison}{function to use for compare}

\item{sampleName}{name of sample to analyze}

\item{subFileExtension}{extension to pre pend to file name}

\item{bonferroniThreshold}{bonferroni threshold to validate pValue}

\item{probeFeatures}{probes features probe, chr, start,end}
}
\value{
list of lesion count  and probes count
}
\description{
Analyze a single sample
}
