% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_annotate_position_pivots.R
\name{anno_annotate_position_pivots}
\alias{anno_annotate_position_pivots}
\title{create an annotated file for each marker, figure, area and subarea, each file has all the sample_groups used to calculate epimutation}
\usage{
anno_annotate_position_pivots()
}
\value{
nothing
}
\description{
create an annotated file for each marker, figure, area and subarea, each file has all the sample_groups used to calculate epimutation
}
