% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_bed.R
\name{annotate_bed}
\alias{annotate_bed}
\title{takes a bed and its location (build with the details of popuationa nd genomic area)
and annoate with detail about genomic area}
\usage{
annotate_bed(
  envir,
  populations,
  figures,
  anomalies,
  groups,
  probes_prefix,
  columnLabel,
  groupingColumnLabel
)
}
\arguments{
\item{envir}{semseekere working infos}

\item{populations}{vector of population to cycle with to build the folder path}

\item{figures}{vector of hyper /hypo to use to build the folder path}

\item{anomalies}{vector of lesions/mutations to use to build the folder path}

\item{groups}{vector of genomic are to cycle and group the annotated data}

\item{probes_prefix}{prefix to use to get the annotated probes dataset}

\item{columnLabel}{label of the column of the genomic area gene, island ,dmr etc..}

\item{groupingColumnLabel}{label of the column of the genomic sub area body, tss1500}
}
\value{
original bed with genomic area infos
}
\description{
takes a bed and its location (build with the details of popuationa nd genomic area)
and annoate with detail about genomic area
}
