% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_bed.R
\name{annotate_bed}
\alias{annotate_bed}
\title{create an annotated file for each marker, figure, area and subarea, each file has all the sample_groups used to calculate epimutation}
\usage{
annotate_bed()
}
\value{
nothing
}
\description{
create an annotated file for each marker, figure, area and subarea, each file has all the sample_groups used to calculate epimutation
}
