% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_analysis.R
\name{association_analysis}
\alias{association_analysis}
\title{Association analysis of SEMseeker's results}
\usage{
association_analysis(inference_details, result_folder, maxResources)
}
\arguments{
\item{inference_details}{independent variable: deve essere nalla sample sheet passata a semseeker
quando lo abbiamo eseguito la prima volta
tipo di regressioni: gaussian, poisson, binomial,quantreg_tau_runs(both as number) eg quantreg_0.25_2000
tipi di test: wilcoxon, stats::t.test,
tipi di correlazioni: pearson, kendall, spearman
MUTATIONS_* ~ tcdd_mother + exam_age
transformation to be applied to dependent variable
(mutations and lesions): scale, log, log2, log10, exp, johnson, none, quantile_quantiles(as number) eg quantile_3
depth analysis:
1: sample level
2: type level (gene, DMR, cpgisland) (includes 1)
3: genomic area: gene, body, gene tss1550, gene whole, gene tss200,  (includes 1 and 2)
filter_p_value report after adjusting saves only significant nominal p-value}

\item{result_folder}{where semseeker's results are stored, the root folder}

\item{maxResources}{percentage of max system's resource to use}
}
\value{

}
\description{
Association analysis of SEMseeker's results
}
