% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_analysis.R
\name{bayes_analysis}
\alias{bayes_analysis}
\title{P_to_be_Case_cond_to_be_Epimutated and sensibility analysis of SEMseeker's mutations and lesions}
\usage{
bayes_analysis(
  result_folder,
  independent_variable = "Sample_Group",
  maxResources = 90,
  parallel_strategy = "multicore",
  ...
)
}
\arguments{
\item{result_folder}{where semseeker's results are stored, the root folder}

\item{maxResources}{percentage of max system's resource to use}

\item{parallel_strategy}{which strategy to use for parallel execution see future vignete: possible values, none, multisession,sequential, multicore, cluster}

\item{...}{other options to filter elaborations}
}
\value{
Invisibly \code{NULL}. Bayesian posterior probability tables
  (\code{bayes_analisys_*.csv}) are written to the \code{Euristic/}
  sub-folder of \code{result_folder}.
}
\description{
P_to_be_Case_cond_to_be_Epimutated and sensibility analysis of SEMseeker's mutations and lesions
}
\examples{
result_dir <- tempdir()
\dontrun{
bayes_analysis(
  result_folder        = "~/semseeker_results/",
  independent_variable = "Sample_Group"
)
}
}
