% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDataSetFromGEO.R
\name{buildDataSetFromGEO}
\alias{buildDataSetFromGEO}
\title{buildDataSetFromGEO}
\usage{
buildDataSetFromGEO(GEOgse, workingFolder, downloadFiles = FALSE)
}
\arguments{
\item{GEOgse}{geo accession dataset identification}

\item{workingFolder}{where sample sheet and files will be saved}

\item{downloadFiles}{true if you wnatr download files from Gene Expression Ombibus (GEO)}
}
\value{
samplesheet, and sample's file saved and samplesheet csv
}
\description{
buildDataSetFromGEO
}
\examples{
buildDataSetFromGEO("GSE186766","~/", TRUE)
}
