% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot}
\usage{
calculate(
  sampleSheetPath,
  methylationData,
  resultFolder,
  bonferroniThreshold = 0.05
)
}
\arguments{
\item{sampleSheetPath}{dataframe with at least a column Sample_ID to identify
samples}

\item{methylationData}{matrix of methylation data}

\item{resultFolder}{folder to save computed epimutations and bedgraphs files.}

\item{bonferroniThreshold}{= 0.05 #threshold to define which pValue
accept for lesions definition}
}
\value{
files into the result folder with pivot table and bedgraph.
}
\description{
Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot
}
