% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heatmap.R
\name{create_heatmap}
\alias{create_heatmap}
\title{create_heatmap load the multiple bed resulting from
analysis organized into files and folders per marker and produce a pivot}
\usage{
create_heatmap()
}
\value{
nothing
}
\description{
create_heatmap load the multiple bed resulting from
analysis organized into files and folders per marker and produce a pivot
}
