% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHeatmap.R
\name{createHeatmap}
\alias{createHeatmap}
\title{createChartFromMultipleBedGenericPerRegion load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot}
\usage{
createHeatmap(
  inputBedDataFrame,
  anomalies,
  groupLabel,
  groupColumnIDs,
  resultFolder,
  logFolder
)
}
\arguments{
\item{anomalies}{vector of anomalies MUTATIONS or LESIONS}

\item{groupLabel}{name of column to group the data set}

\item{resultFolder}{folder as root for bedfiles organized per anomaly}

\item{logFolder}{log folder used by the parallel cluster}

\item{populations}{vectors of populations name used to group populations
for differential analysis}

\item{figures}{vectors of figure to identify lesions as HYPER or HYPO}

\item{subGroups}{other grouping column over main group}

\item{probesPrefix}{definition of prefix to use probes to load}

\item{subGroupLabel}{name of column to sub group the data set}
}
\value{
list of pivot by column identified with groupLabel and by Sample
}
\description{
createChartFromMultipleBedGenericPerRegion load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot
}
