% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHeatmap.R
\name{createHeatmap}
\alias{createHeatmap}
\title{createHeatmap load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot}
\usage{
createHeatmap(envir, inputBedDataFrame, anomalies, groupLabel, groupColumnIDs)
}
\arguments{
\item{envir}{semseekere working infos}

\item{inputBedDataFrame}{data frame to chart}

\item{anomalies}{vector of anomalies to manage}

\item{groupLabel}{main genomic area to char eg: gene}

\item{groupColumnIDs}{positions of the group coplumn id}
}
\value{
list of pivot by column identified with groupLabel and by Sample
}
\description{
createHeatmap load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot
}
