% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPivotResultFromMultipleBedGeneric.R
\name{createPivotResultFromMultipleBedGeneric}
\alias{createPivotResultFromMultipleBedGeneric}
\title{createPivotResultFromMultipleBedGeneric load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot}
\usage{
createPivotResultFromMultipleBedGeneric(
  genomicAreaMain,
  genomicAreaSub,
  sampleSheet
)
}
\arguments{
\item{anomalyLabel}{anomaly definition used to lable folder and files eg
MUTATIONS, LESIONS}

\item{figureLable}{figure used to create the file HYPO HYPER}

\item{probeFeatures}{features of probe CHR and START and NAME}

\item{columnLabel}{}
}
\value{
list of pivot by column identified with columnLabel and by Sample
}
\description{
createPivotResultFromMultipleBedGeneric load the multiple bed resulting from
analysis organized into files and folders per anomaly and produce a pivot
}
