% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummaryExcelFromCumulativeBedFile.R
\name{createSummaryExcelFromCumulativeBedFile}
\alias{createSummaryExcelFromCumulativeBedFile}
\title{createSummaryExcelFromCumulativeBedFile build an excel summary of figures and
anomaly staring from the cumulative bed files into the result folder}
\usage{
createSummaryExcelFromCumulativeBedFile(probeFeatures, sampleSheet)
}
\arguments{
\item{probeFeatures}{probes information CHR, START, END and rowname
as probenames}

\item{sampleSheet}{sheet of samples with at least Sample_ID column}
}
\value{
a saved excel file
}
\description{
anomaly and figure built tree
}
