% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{data_preparation}
\alias{data_preparation}
\title{Title}
\usage{
data_preparation(
  family_test,
  transformation,
  tempDataFrame,
  independent_variable,
  g_start,
  dototal,
  covariates,
  depth_analysis,
  envir
)
}
\arguments{
\item{envir}{}
}
\description{
Title
}
