% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltar_single_sample.R
\name{deltar_single_sample}
\alias{deltar_single_sample}
\title{delta_single_sample}
\usage{
deltar_single_sample(values, thresholds, sample_detail)
}
\arguments{
\item{values}{values of methylation}

\item{sample_detail}{details of sample to analyze}

\item{high_thresholds}{highest threshold to use for comparison}

\item{low_thresholds}{lowest threshold to use for comparison}

\item{signal_medians}{median to use for calculation}

\item{probe_features}{genomic position of probe_features}
}
\value{
summary detail about the analysis
}
\description{
delta_single_sample
}
