% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltar_get.R
\name{deltar_single_sample}
\alias{deltar_single_sample}
\title{deltar_single_sample}
\usage{
deltar_single_sample(
  values,
  high_thresholds,
  low_thresholds,
  sample_detail,
  iqr,
  probe_features
)
}
\arguments{
\item{values}{values of methylation}

\item{high_thresholds}{highest threshold to use for comparison}

\item{low_thresholds}{lowest threshold to use for comparison}

\item{sample_detail}{details of sample to analyze}

\item{iqr}{interqurtile ratio used to calculate the beta range}

\item{probe_features}{genomic position of probes}
}
\value{
summary detail about the analysis
}
\description{
deltar_single_sample
}
