% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaSingleSample.R
\name{deltaSingleSample}
\alias{deltaSingleSample}
\title{deltaSingleSample}
\usage{
deltaSingleSample(
  values,
  highThresholds,
  lowThresholds,
  sampleDetail,
  betaMedians,
  probeFeatures
)
}
\arguments{
\item{values}{values of methylation}

\item{highThresholds}{highest threshold to use for comparison}

\item{lowThresholds}{lowest threshold to use for comparison}

\item{sampleDetail}{details of sample to analyze}

\item{betaMedians}{median to use for calculation}

\item{probeFeatures}{}
}
\value{
none
}
\description{
deltaSingleSample
}
