% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaSingleSample.R
\name{deltaSingleSample}
\alias{deltaSingleSample}
\title{deltaSingleSample}
\usage{
deltaSingleSample(
  values,
  resultFolder,
  highThresholds,
  lowThresholds,
  sampleName,
  betaMedians,
  subFileExtension,
  probeFeatures
)
}
\arguments{
\item{values}{values of methylation}

\item{resultFolder}{folder to store result files}

\item{highThresholds}{highest threshold to use for comparison}

\item{lowThresholds}{lowest threshold to use for comparison}

\item{sampleName}{name of sample to analyze}

\item{betaMedians}{median to use for calculation}

\item{subFileExtension}{extension to pre pend to file name}
}
\value{
none
}
\description{
deltaSingleSample
}
