% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumpSampleAsBedFile.R
\name{dumpSampleAsBedFile}
\alias{dumpSampleAsBedFile}
\title{Title}
\usage{
dumpSampleAsBedFile(
  dataToDump,
  fileExtension,
  resultFolder,
  resultSubFolder,
  sampleName,
  multipleFileColNames
)
}
\arguments{
\item{multipleFileColNames}{}
}
\value{

}
\description{
Title
}
