% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLesions.R
\name{getLesions}
\alias{getLesions}
\title{Title}
\usage{
getLesions(
  mutationAnnotatedSorted,
  slidingWindowSize,
  sampleName,
  probePositions,
  bonferroniThreshold = 0.05
)
}
\arguments{
\item{bonferroniThreshold}{}
}
\value{

}
\description{
Title
}
