% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutations.R
\name{getMutations}
\alias{getMutations}
\title{getMutations}
\usage{
getMutations(values, figure, thresholds, probeFeatures, sampleName)
}
\arguments{
\item{values}{values of methylation}

\item{figure}{figure to get Mutaions of HYPO or HYPER methylation}

\item{thresholds}{threshold to use for comparison}

\item{probeFeatures}{probes features probe, chr, start,end}

\item{sampleName}{name of the sample}
}
\value{
mutations
}
\description{
getMutations
}
