% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_model.R
\name{glm_model}
\alias{glm_model}
\title{Title}
\usage{
glm_model(family_test, tempDataFrame, sig.formula)
}
\arguments{
\item{family_test}{regression model to apply}

\item{tempDataFrame}{data frame to use for the model}

\item{sig.formula}{formula to apply the model}
}
\description{
Title
}
