% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_env.R
\name{init_env}
\alias{init_env}
\title{init environment}
\usage{
init_env(resultFolder, maxResources = 90)
}
\arguments{
\item{resultFolder}{where result of semseeker will bestored}

\item{maxResources}{percentage of how many available cores will be used default 90 percent, rounded to the lowest integer}
}
\value{

}
\description{
init environment
}
