% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_study_association.R
\name{intra_study_association}
\alias{intra_study_association}
\title{Intra Study Association (deprecated alias)}
\usage{
intra_study_association(...)
}
\arguments{
\item{...}{Argomenti inoltrati a `intra_study_association_replication()`.}
}
\value{
Invisibly `NULL`. Vedi [intra_study_association_replication()].
}
\description{
Alias deprecato di [intra_study_association_replication()]. Naming uniformato a
  `<scope>_<object>_<operation>` (es. `intra_study_*`,
  `inter_study_*`). Sostituire le chiamate `intra_study_association(...)` con
  `intra_study_association_replication(...)` (stessa firma). Alla chiamata viene emesso un warning
  di deprecation.
}
\seealso{
[intra_study_association_replication()]
}
