% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot_marker_per_sample.R
\name{manhattan_plot_marker_per_sample}
\alias{manhattan_plot_marker_per_sample}
\title{Title manhattan_plot_marker_per_sample}
\usage{
manhattan_plot_marker_per_sample(
  sample_name = "NAME",
  probes_range = 1000:2000,
  hyper_color = "blue",
  hypo_color = "orange",
  non_outlier_color = "grey",
  limit_label_color = ssEnv$color_palette[1],
  result_folder = result_folder,
  maxResources = maxResources,
  parallel_strategy = parallel_strategy,
  ...
)
}
\arguments{
\item{sample_name}{name of the sample to be plotted}

\item{probes_range}{range of probes to be plotted}

\item{hyper_color}{color for hypermethylated probes}

\item{hypo_color}{color for hypomethylated probes}

\item{non_outlier_color}{color for non-outlier probes}

\item{limit_label_color}{color for limit labels}

\item{result_folder}{folder where the results are stored}

\item{maxResources}{maximum number of resources to be used}

\item{parallel_strategy}{strategy to be used for parallelization}

\item{...}{}
}
\description{
Title manhattan_plot_marker_per_sample
}
