% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot_per_area.R
\name{manhattan_plot_per_area}
\alias{manhattan_plot_per_area}
\title{Title}
\usage{
manhattan_plot_per_area(
  marker,
  figure,
  group,
  subgroup,
  family,
  adjust_method,
  phenotype,
  only_significant_areas = FALSE
)
}
\arguments{
\item{marker}{investigated marker eg. MUTATIONS, DELTAR, DELTAQ}

\item{figure}{HYPO, HYPER}

\item{group}{genomic area (eg. GENE, ISLAND, DMR)}

\item{subgroup}{sub genomic area (TSS1550), depending on the genomic area}

\item{family}{fullname of the family used for the association analysis}

\item{adjust_method}{colnames of the pvalue adjusted to use}

\item{phenotype}{variable to select from the sample_sheet to use for coloring point}

\item{only_significant_areas}{TRUE if filter for pvalue < 0.05}
}
\description{
Title
}
