% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_performance_association_q_search.R
\name{markers_performance_association_q_search}
\alias{markers_performance_association_q_search}
\title{calculate scores for th e same dataset at changing of quantile/bins}
\usage{
markers_performance_association_q_search(
  inference_details,
  result_folder,
  dest_folder,
  folder_key,
  pvalue_column = "PVALUE_ADJ_ALL_BH",
  aggr_fun = "mean",
  significance = TRUE,
  sql_conditions = c(),
  study,
  qs = c(2, 4, 8, 16, 32),
  ...
)
}
\description{
calculate scores for th e same dataset at changing of quantile/bins
}
