% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_permutation_model.R
\name{mean_permutation}
\alias{mean_permutation}
\title{Title}
\usage{
mean_permutation(
  family_test,
  sig.formula,
  tempDataFrame,
  independent_variable,
  plot,
  samples_sql_condition = samples_sql_condition,
  key
)
}
\arguments{
\item{family_test}{family lqmm, mean}

\item{sig.formula}{formula of the model}

\item{tempDataFrame}{data}

\item{independent_variable}{name of regressor}

\item{permutation_success}{number of success tests to calculate corrected confidence interval}

\item{tests_count}{count of total executed tests}
}
\value{
A numeric p-value from the permutation-based mean-difference test.
}
\description{
Title
}
