% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations_get.R
\name{mutations_get}
\alias{mutations_get}
\title{mutations_get}
\usage{
mutations_get(values, figure, thresholds, sampleName)
}
\arguments{
\item{values}{values of methylation}

\item{figure}{figure to get Mutaions of HYPO or HYPER methylation}

\item{thresholds}{threshold to use for comparison}

\item{sampleName}{name of the sample}
}
\value{
mutations
}
\description{
mutations_get
}
