% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_to_long_format.R
\name{pivot_to_long_format}
\alias{pivot_to_long_format}
\title{Get the pivot in long format instead of wide format}
\usage{
pivot_to_long_format(
  marker,
  figure,
  group,
  subgroup,
  phenotype_column,
  sample_sheet,
  areas_selection = NULL
)
}
\arguments{
\item{marker}{marker to filer HYPER, HYPO, BOTH}

\item{figure}{DELTAS, DELTAQ,DELTAR, MUTATIONS}

\item{group}{GENE, DMR ...}

\item{subgroup}{TSS1500 ...}

\item{phenotype_column}{column from the sample sheet to pair to each sample}

\item{sample_sheet}{sample sheet of samples}

\item{areas_selection}{genomic area to select, if NULL all areas will be selected}
}
\value{
the pivot in a long format of 3 columnns, the phontype column with name phenotype, the value of the marker and the area investigated
}
\description{
Get the pivot in long format instead of wide format
}
