% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_permutation_model.R
\name{quantile_permutation_model}
\alias{quantile_permutation_model}
\title{quantile_permutation_model calculate differences between the same quantile of two distribution}
\usage{
quantile_permutation_model(
  family_test,
  sig.formula,
  tempDataFrame,
  independent_variable,
  transformation,
  plot = FALSE
)
}
\arguments{
\item{family_test}{family quantile}

\item{sig.formula}{formula of the model}

\item{tempDataFrame}{data}

\item{independent_variable}{name of regressor}

\item{permutation_success}{number of success tests to calculate corrected confidence interval}

\item{tests_count}{count of total executed tests}
}
\description{
quantile_permutation_model calculate differences between the same quantile of two distribution
}
