% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantreg_model.R
\name{quantreg_model}
\alias{quantreg_model}
\title{Title}
\usage{
quantreg_model(
  family_test,
  sig.formula,
  tempDataFrame,
  independent_variable,
  transformation,
  plot
)
}
\arguments{
\item{family_test}{family lqmm, quantreg}

\item{sig.formula}{formula of the model}

\item{tempDataFrame}{data}

\item{independent_variable}{name of regressor}

\item{permutation_success}{number of success tests to calculate corrected confidence interval}

\item{tests_count}{count of total executed tests}
}
\description{
Title
}
