% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantreg_model.R
\name{quantreg_model}
\alias{quantreg_model}
\title{Title}
\usage{
quantreg_model(
  family_test,
  sig.formula,
  tempDataFrame,
  independent_variable,
  boot_success,
  tests_count
)
}
\arguments{
\item{tests_count}{}
}
\description{
Title
}
