% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMultipleBed.R
\name{readMultipleBed}
\alias{readMultipleBed}
\title{read multiple bed with annotated data as per input parameter}
\usage{
readMultipleBed(
  envir,
  anomalyLabel,
  figureLable,
  probe_features,
  columnLabel,
  populationName,
  groupingColumnLabel
)
}
\arguments{
\item{envir}{semseekere working infos}

\item{anomalyLabel}{anomaly definition used to label folder and files eg MUTATIONS, LESIONS}

\item{figureLable}{figures like hypo/hyper to built the data path}

\item{probe_features}{features of probe CHR and START and NAME}

\item{columnLabel}{name of column in the annotation dataset to select genomic area (gene, island etc..)}

\item{populationName}{name of the population used to build the data path}

\item{groupingColumnLabel}{name of the genomic sub area}
}
\value{
list of pivot by column identified with column Label and by Sample
}
\description{
read multiple bed with annotated data as per input parameter
}
