% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_signal_single_sample.R
\name{sem_signal_single_sample}
\alias{sem_signal_single_sample}
\title{sem_signal_single_sample}
\usage{
sem_signal_single_sample(values, sample_detail, probe_features)
}
\arguments{
\item{values}{signal vaòues}

\item{sample_detail}{detais of sample}

\item{probe_features}{annotation probe}
}
\value{
signal mean
}
\description{
sem_signal_single_sample
}
