% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semseeker.R
\name{semseeker}
\alias{semseeker}
\title{Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot}
\usage{
semseeker(sample_sheet, signal_data, result_folder, ...)
}
\arguments{
\item{sample_sheet}{dataframe with at least a column Sample_ID to identify samples}

\item{signal_data}{matrix of methylation data}

\item{result_folder}{where the result will be saved}

\item{...}{other options to execute elaborations}
}
\value{
files into the result folder with pivot table and bedgraph.
}
\description{
Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot
}
