% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semseeker.R
\name{semseeker}
\alias{semseeker}
\title{Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot}
\usage{
semseeker(
  sampleSheet,
  methylationData,
  resultFolder,
  bonferroniThreshold = 0.05,
  inferenceDetails = NULL
)
}
\arguments{
\item{sampleSheet}{dataframe with at least a column Sample_ID to identify
samples}

\item{methylationData}{matrix of methylation data}

\item{bonferroniThreshold}{= 0.05 #threshold to define which pValue}

\item{inferenceDetails}{dataframe of details to calculate inferential statistics
accept for lesions definition}
}
\value{
files into the result folder with pivot table and bedgraph.
}
\description{
Calculate stochastic epi mutations from a methylation dataset as outcome
report of pivot
}
