% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_range_values.R
\name{signal_range_values}
\alias{signal_range_values}
\title{calculate the range of signal values to define the outlier}
\usage{
signal_range_values(populationMatrix)
}
\arguments{
\item{populationMatrix}{matrix of methylation for the population under calculation}
}
\value{
methylation matrix as normalized distribution
}
\description{
calculate the range of signal values to define the outlier
}
