% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_analysis.R
\name{ss_analysis}
\alias{ss_analysis}
\title{sensitivity and sensibility analysis of SEMseeker's mutations and lesions}
\usage{
ss_analysis(
  samples_sql_selection = "",
  combinations,
  result_folder,
  independent_variable = "Sample_Group",
  maxResources = 90,
  parallel_strategy = "multicore",
  ...
)
}
\arguments{
\item{result_folder}{where semseeker's results are stored, the root folder}

\item{maxResources}{percentage of max system's resource to use}

\item{parallel_strategy}{which strategy to use for parallel execution see future vignete: possible values, none, multisession,sequential, multicore, cluster}

\item{...}{other options to filter elaborations}
}
\value{
Invisibly \code{NULL}. Sensitivity / specificity tables are written
  to the \code{Euristic/} sub-folder of \code{result_folder}.
}
\description{
sensitivity and sensibility analysis of SEMseeker's mutations and lesions
}
\examples{
result_dir <- tempdir()
\donttest{
ss_analysis(
  combinations         = list(c("CASE_A", "CASE_B")),
  result_folder        = "~/semseeker_results/",
  independent_variable = "Sample_Group"
)
}
}
