% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_analysis.R
\name{ss_analysis}
\alias{ss_analysis}
\title{sensitivity and sensibility analysis of SEMseeker's mutations and lesions}
\usage{
ss_analysis(
  result_folder,
  independent_variable = "Sample_Group",
  maxResources = 90,
  parallel_strategy = "multicore",
  ...
)
}
\arguments{
\item{result_folder}{where semseeker's results are stored, the root folder}

\item{maxResources}{percentage of max system's resource to use}

\item{parallel_strategy}{which strategy to use for parallel execution see future vignete: possible values, none, multisession,sequential, multicore, cluster}

\item{...}{other options to filter elaborations}
}
\description{
sensitivity and sensibility analysis of SEMseeker's mutations and lesions
}
