% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model.R
\name{test_model}
\alias{test_model}
\title{Title}
\usage{
test_model(
  family_test,
  tempDataFrame,
  sig.formula,
  burdenValue,
  independent_variable
)
}
\arguments{
\item{family_test}{which family test to apply}

\item{tempDataFrame}{data frame to use with the test}

\item{sig.formula}{formula to apply}

\item{burdenValue}{burden colon name}

\item{independent_variable}{independent variable for regressor}
}
\description{
Title
}
