% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model.R
\name{test_model}
\alias{test_model}
\title{Title}
\usage{
test_model(
  family_test,
  tempDataFrame,
  sig.formula,
  burdenValue,
  independent_variable
)
}
\arguments{
\item{independent_variable}{}
}
\description{
Title
}
